INTEROP: Error out when switching function settings mismatch
