+++ openff/interchange/exceptions.py
@@ -106,6 +106,10 @@ class CutoffMismatchError(UnsupportedCombinationError):
     """Non-bonded cutoffs do not match."""
 
 
+class SwitchingFunctionMismatchError(UnsupportedCombinationError):
+    """Switching distances or the use of a switching function does not match."""
+
+
 class PluginCompatibilityError(InterchangeException):
     """A plugin is incompatible with the current version of Interchange in the way it is called."""
